refactored, db script, doc outline
please better document
