--- src/main/java/com/example/demo/AsyncExceptionHandler.java
@@ -1,17 +1,0 @@
-package com.example.demo;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
-
-import java.lang.reflect.Method;
-
-public class AsyncExceptionHandler implements AsyncUncaughtExceptionHandler {
-//    private final Logger logger = LoggerFactory.getLogger(AsyncExceptionHandler.class);
-
-    @Override
-    public void handleUncaughtException(Throwable ex, Method method, Object... params) {
-        System.out.println("Unexpected asynchronous exception at : "
-                + method.getDeclaringClass().getName() + "." + method.getName()+" <--> "+ex);
-    }
-}
