+++ src/main/java/com/example/demo/DemoApplication.java
@@ -2,10 +2,8 @@ package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
-@EnableAsync
 public class DemoApplication {
 
 	public static void main(String[] args) {
