--- src/main/java/com/example/demo/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo;
-
-import com.example.demo.model.Subscribers;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<Subscribers, Integer> {
-
-}
