+++ src/main/java/com/example/demo/subscriber/Subscriber.java
@@ -1,33 +1,33 @@
-package com.example.demo.model;
+package com.example.demo.subscriber;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
 import org.hibernate.annotations.Generated;
 import org.hibernate.annotations.GenerationTime;
 
-import javax.persistence.*;
-
 @Entity
-@Table(name="spam_ctrl_db8")
-public class Subscribers {
+@Table(name = "spam_ctrl_db8")
+public class Subscriber {
     @Id
-    @Column(name = "msisdn")
     private int msisdn;
-    @Column(name = "mode")
+
     private String mode;
 
-    @Column(name = "timestamp")
+    // while "timestamp" is no "reserved keyword" in postgres, it is in standard SQL
+    // : https://www.postgresql.org/docs/current/sql-keywords-appendix.html
     @Generated(value = GenerationTime.INSERT)
     private String timestamp;
 
-    public Subscribers() {
+    public Subscriber() {
         // default constructor
     }
 
-
-    public Subscribers(int msisdn, String mode) {
+    public Subscriber(int msisdn, String mode) {
         this.msisdn = msisdn;
         this.mode = mode;
     }
-
 
     public int getMsisdn() {
         return msisdn;
