+++ src/main/java/com/example/demo/subscriber/SubscriberRepository.java
@@ -1,0 +1,7 @@
+package com.example.demo.subscriber;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SubscriberRepository extends JpaRepository<Subscriber, Integer> {
+
+}
