+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.datasource.username=deepak
 spring.datasource.password=deepak
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
+spring.jpa.open-in-view=false
