+++ src/test/java/com/example/demo/DemoApplicationTests.java
@@ -1,13 +1,17 @@
 package com.example.demo;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.ApplicationContext;
 
 @SpringBootTest
 class DemoApplicationTests {
 
 	@Test
-	void contextLoads() {
+	void contextLoads(ApplicationContext ctxt) {
+		assertNotNull(ctxt);
 	}
 
 }
